separate date picker to date time and date range
