+++ app/javascript/dashboard/routes/dashboard/settings/inbox/components/AddCampaign.vue
@@ -32,8 +32,7 @@
 
         <label v-if="isOnOffType">
           {{ $t('CAMPAIGN.ADD.FORM.SCHEDULED_AT.LABEL') }}
-          <woot-date-picker
-            type="datetime"
+          <woot-date-time-picker
             :value="scheduledAt"
             :confirm-text="$t('CAMPAIGN.ADD.FORM.SCHEDULED_AT.CONFIRM')"
             :placeholder="$t('CAMPAIGN.ADD.FORM.SCHEDULED_AT.PLACEHOLDER')"
@@ -143,10 +142,10 @@ import { mapGetters } from 'vuex';
 import { required, url, minLength } from 'vuelidate/lib/validators';
 import alertMixin from 'shared/mixins/alertMixin';
 import campaignMixin from 'shared/mixins/campaignMixin';
-import WootDatePicker from 'dashboard/components/ui/DatePicker.vue';
+import WootDateTimePicker from 'dashboard/components/ui/DateTimePicker.vue';
 
 export default {
-  components: { WootDatePicker },
+  components: { WootDateTimePicker },
   mixins: [alertMixin, campaignMixin],
   props: {
     senderList: {
